feat: introduce full per-peer task accounting and truly async connection handling
