+++ network/Cargo.toml
@@ -69,6 +69,9 @@ version = "0.4.2"
 [dependencies.log]
 version = "0.4.11"
 
+[dependencies.once_cell]
+version = "1.5.2"
+
 [dependencies.parking_lot]
 version = "0.11.1"
