+++ testing/src/network/mod.rs
@@ -158,7 +158,6 @@ pub fn test_consensus(setup: ConsensusSetup, node: Node<LedgerStorage>) -> Conse
 /// Returns an `Environment` struct with given arguments
 pub fn test_environment(setup: TestSetup) -> Environment {
     Environment::new(
-        setup.socket_address,
         setup.min_peers,
         setup.max_peers,
         setup.bootnodes,
@@ -179,7 +178,8 @@ pub async fn test_node(setup: TestSetup) -> Node<LedgerStorage> {
         node.set_consensus(consensus);
     }
 
-    node.start().await.unwrap();
+    node.listen(setup.socket_address).await.unwrap();
+    node.start_services().await;
 
     if is_miner {
         let miner_address = FIXTURE.test_accounts[0].address.clone();
